api/requirements.txt
# Core Flask dependencies
Flask==2.2.3
flask-cors==6.0.1
Werkzeug==2.3.8
python-dotenv==1.1.1

# Azure Form Recognizer
azure-ai-formrecognizer==3.3.3
azure-core==1.35.0

# Machine Learning and AI
ultralytics==8.3.155
torch==2.3.0
torchvision==0.18.0
scikit-learn==1.6.1
joblib==1.5.1

# Data processing
numpy==1.26.4
pandas==2.1.4
opencv-python==4.8.1.78
Pillow==10.0.1

# Utilities
requests==2.32.4